- include stdint.h - optimized int type - deleted warning in function_writter

+++ src/asm/lexer/file.c
@@ -9,15 +9,15 @@
 
 static char *get_file_name(char *str)
 {
-	int path = 0;
-	int len = 0;
+	uint8_t path = 0;
+	uint8_t len = 0;
 
-	for (int i = 0; str[i] != '\0'; i++)
+	for (uint8_t i = 0; str[i] != '\0'; i++)
 		if (str[i] == '/')
 			path = i + 1;
-	for (int j = path; str[j] != '.' && str[j] != '\0'; j++)
+	for (uint8_t j = path; str[j] != '.' && str[j] != '\0'; j++)
 		len += 1;
-	for (int k = 0; k < len; k++)
+	for (uint8_t k = 0; k < len; k++)
 		str[k] = str[path++];
 
 	str[len] = '\0';
@@ -50,7 +50,7 @@ static char *open_file(char *file)
 		free(str);
 		return (NULL);
 	}
-	for (int i = 1; (size = read(fd, buff, 1)) != 0; i++) {
+	for (uint8_t i = 1; (size = read(fd, buff, 1)) != 0; i++) {
 		if (size == -1)
 			return (NULL);
 		str = realloc(str, i + 1);
